refactor: consolidate access service into single impl.go file
Merge stub.go methods into impl.go with clear section delineation:
- Grants (low-level principal_grants CRUD)
- Roles & Members (WorkOS-backed role management)
- Utilities (slugify)
